+++ src/constants/index.ts
@@ -1,1 +1,4 @@
+/** it's means 5 minute only */
+export const MAIL_EXPIRE_TIME = 5 as const;
 export * from './enums';
+