+++ changelog.md
@@ -1,5 +1,6 @@
 ## v0.0.1 🎉
 
 - (Auth Module) Login and and signup with bearer token
+  - login, logout, forget password, resent code, reset password
 - (Post Module) Creating a new post with author id
 - (Category Module) Creating a new category for assign the category id to the post (optional)